+++ src/game/server/tf/nav_mesh/tf_nav_mesh.cpp
@@ -2665,3 +2665,18 @@ void CTFNavMesh::CollectAreaWithinBombTravelRange( CUtlVector< CTFNavArea * > *s
 		}
 	}
 }
+
+void CTFNavMesh::ScriptRecalculateBlocked( float delay )
+{
+	ScheduleRecomputationOfInternalData( MAP_LOGIC, delay );
+}
+
+void CTFNavMesh::ScriptRecalculateBlockedWithCapture( float delay )
+{
+	ScheduleRecomputationOfInternalData( POINT_CAPTURED, delay );
+}
+
+BEGIN_SCRIPTDESC( CTFNavMesh, CNavMesh, "The TF nav mesh" )
+DEFINE_SCRIPTFUNC_NAMED( ScriptRecalculateBlocked, "RecomputeBlockers", "Arguments: ( delay ) - schedule recalculating blocked state of all areas" )
+DEFINE_SCRIPTFUNC_NAMED( ScriptRecalculateBlockedWithCapture, "RecomputeBlockersWithCapture", "Arguments: ( delay ) - schedule recalculating blocked state of all areas, include point capture blocking" )
+END_SCRIPTDESC();